Bloomfilter
For this subtask, you must implement a bloom filter in order to find rgbs in the stream that match a given set of colors, determined by the first 2000 elements.  However, because no one can perceive small differences in million-value RGB codes, two triples of rgb values will be considered a match if each of the r, g, b values differ by [updated: 2/21]: 1 or less. That is, two triples (r1, g1, b1) and (r2, g2, b2) are considered a match iff ((abs(r1-r2) <= 1) and (abs(g1-g2) <= 1) and (abs(b1-b2) < 1)).   Note, you will need to come up with a hashing algorithm or approach that is able to handle this matching criterion and keep your false positives on the test set to under 1%.
You will need to implement two methods:
task1b_bloomSetup(elements_in_set)
task1b_bloomStream(element)
bloomSetup implements the setup of a Bloom filter where by the bit array is created. bloomStream is responsible for processing an element at a time, returning true if it's in the filter and false otherwise.  Besides the 1000 element array, you may store any number of hash functions (considered part of the code).
[added: 2/21]:
extra credit (+10): implement a version of the bloom filter that allows for a difference in rgb values up to 100.  That is, two triples (r1, g1, b1) and (r2, g2, b2) are considered a match iff ((abs(r1-r2) < 100) and (abs(g1-g2) <= 100) and (abs(b1-b2) < 100)).